log-generator/requirements.txt
psycopg2-binary
black
dotenv